Split system adaptor into it's own class
