+++ src/ML-METATENSOR/pair_metatensor.cpp
@@ -17,7 +17,6 @@
 #include "pair_metatensor.h"
 
 #include "atom.h"
-#include "domain.h"
 #include "error.h"
 #include "force.h"
 #include "memory.h"
@@ -27,25 +26,25 @@
 #include "comm.h"
 
 #include "neigh_list.h"
-#include "neigh_request.h"
 
 #include <torch/script.h>
 
 #include <metatensor/torch.hpp>
 #include <metatensor/torch/atomistic.hpp>
 
-using namespace LAMMPS_NS;
+#include "metatensor_system.h"
 
-static metatensor_torch::System system_from_lmp(Atom* atom);
+using namespace LAMMPS_NS;
 
 /* ---------------------------------------------------------------------- */
 
-PairMetatensor::PairMetatensor(LAMMPS *lmp) : Pair(lmp) {
+PairMetatensor::PairMetatensor(LAMMPS *lmp):
+    Pair(lmp),
+    system_adaptor(nullptr)
+{
     auto options = torch::TensorOptions().dtype(torch::kInt32);
     this->selected_atoms_values = torch::zeros({0, 2}, options);
-    this->atomic_types = torch::zeros({0}, options);
 
-    this->interaction_range = -1;
     // default to true for now, this will be changed to false later
     this->check_consistency = true;
 
@@ -82,14 +81,12 @@ PairMetatensor::PairMetatensor(LAMMPS *lmp) : Pair(lmp) {
 }
 
 PairMetatensor::~PairMetatensor() {
-    delete this->lammps_to_metatensor_types;
-
     if (allocated) {
         memory->destroy(setflag);
         memory->destroy(cutsq);
+        memory->destroy(type_mapping);
     }
 }
-
 
 // called when finding `pair_style metatensor` in the input
 void PairMetatensor::settings(int argc, char ** argv) {
@@ -149,15 +146,19 @@ void PairMetatensor::allocate() {
 
     // lammps_types_to_species stores the mapping from lammps atom types to
     // the metatensor model species
-    lammps_to_metatensor_types = memory->create(
-        lammps_to_metatensor_types,
+    type_mapping = memory->create(
+        type_mapping,
         atom->ntypes + 1,
-        "PairMetatensor:lammps_types_to_species"
+        "PairMetatensor:type_mapping"
     );
 
     for (int i = 1; i <= atom->ntypes; i++) {
-        lammps_to_metatensor_types[i] = -1;
+        type_mapping[i] = -1;
     }
+}
+
+double PairMetatensor::init_one(int, int) {
+    return this->interaction_range;
 }
 
 
@@ -175,8 +176,8 @@ void PairMetatensor::coeff(int argc, char ** argv) {
     }
 
     for (int lammps_type=1; lammps_type<argc - 1; lammps_type++) {
-        int species = utils::inumeric(FLERR, argv[lammps_type + 1], true, lmp);
-        lammps_to_metatensor_types[lammps_type] = species;
+        int type = utils::inumeric(FLERR, argv[lammps_type + 1], true, lmp);
+        type_mapping[lammps_type] = type;
     }
 
     // mark all pairs coeffs as known
@@ -199,6 +200,24 @@ void PairMetatensor::init_style() {
         error->all(FLERR, "Pair style metatensor requires newton pair on");
     }
 
+    // get the model's interaction range
+    auto range = this->capabilities->engine_interaction_range(this->evaluation_options->length_unit());
+    if (range < 0) {
+        error->all(FLERR, "interaction_range is negative for this model");
+    } else if (!std::isfinite(range)) {
+        error->all(FLERR, "interaction_range is infinite for this model, this is not yet supported");
+    } else {
+        this->interaction_range = range;
+    }
+
+    // create system adaptor
+    auto options = MetatensorSystemOptions{
+        type_mapping,
+        this->interaction_range,
+        check_consistency,
+    };
+    this->system_adaptor = std::make_unique<MetatensorSystemAdaptor>(lmp, this, options);
+
     // Translate from the metatensor neighbors lists requests to LAMMPS
     // neighbors lists requests.
     auto requested_nl = this->torch_model->run_method("requested_neighbors_lists");
@@ -206,47 +225,13 @@ void PairMetatensor::init_style() {
         auto options = ivalue.get().toCustomClass<metatensor_torch::NeighborsListOptionsHolder>();
         auto cutoff = options->engine_cutoff(this->evaluation_options->length_unit());
 
-        if (cutoff > this->interaction_range) {
-            error->all(FLERR,
-                "Invalid metatensor model: one of the requested neighbor lists "
-                "has a cutoff ({}) larger than the model interaction range ({})",
-                cutoff, this->interaction_range
-            );
-        } else if (cutoff < 0 || !std::isfinite(cutoff)) {
-            error->all(FLERR,
-                "model requested an invalid cutoff for neighbors list: {} "
-                "(cutoff in model units is {})",
-                cutoff, options->cutoff()
-            );
-        }
-
-        this->neigh_cache.emplace_back(NeighborsData{
-            cutoff,
-            options,
-            /* known_samples */ {},
-            /* samples */ {},
-            /* distances */ {},
-        });
-
+        this->system_adaptor->add_nl_request(cutoff, options);
     }
-
-    // We ask LAMMPS for a full neighbor lists because we need to know about
-    // ALL pairs, even if options->full_list() is false. We will then filter
-    // the pairs to only include each pair once where needed.
-    auto request = neighbor->add_request(this, NeighConst::REQ_FULL | NeighConst::REQ_GHOST);
-    request->set_id(0);
-    request->set_cutoff(this->interaction_range);
 }
 
 
 void PairMetatensor::init_list(int id, NeighList *ptr) {
-    assert(id == 0);
-    this->list = ptr;
-}
-
-
-double PairMetatensor::init_one(int /*i*/, int /*j*/) {
-    return this->interaction_range;
+    system_adaptor->init_list(id, ptr);
 }
 
 
@@ -263,8 +248,17 @@ void PairMetatensor::compute(int eflag, int vflag) {
         this->evaluation_options->outputs.at("energy")->per_atom = false;
     }
 
+    auto dtype = torch::kFloat64;
+    if (this->capabilities->dtype() == "float64") {
+        dtype = torch::kFloat64;
+    } else if (this->capabilities->dtype() == "float32") {
+        dtype = torch::kFloat64;
+    } else {
+        error->all(FLERR, "the model requested an unsupported dtype '{}'", this->capabilities->dtype());
+    }
+
     // transform from LAMMPS to metatensor System
-    auto system = this->system_from_lmp(static_cast<bool>(vflag_global));
+    auto system = system_adaptor->system_from_lmp(static_cast<bool>(vflag_global), dtype);
 
     // only run the calculation for atoms actually in the current domain
     selected_atoms_values.resize_({atom->nlocal, 2});
@@ -325,7 +319,7 @@ void PairMetatensor::compute(int eflag, int vflag) {
     assert(!vflag_fdotr);
 
     if (vflag_global) {
-        auto virial_tensor = this->strain.grad();
+        auto virial_tensor = system_adaptor->strain.grad();
         auto predicted_virial = virial_tensor.accessor<double, 2>();
 
         virial[0] += predicted_virial[0][0];
@@ -359,18 +353,7 @@ void PairMetatensor::load_torch_model(const char* path) {
     }
 
     auto capabilities_ivalue = this->torch_model->run_method("capabilities");
-
     this->capabilities = capabilities_ivalue.toCustomClass<metatensor_torch::ModelCapabilitiesHolder>();
-
-    // get the model's interaction range
-    auto range = this->capabilities->engine_interaction_range(this->evaluation_options->length_unit());
-    if (range < 0) {
-        error->all(FLERR, "interaction_range is negative for this model");
-    } else if (!std::isfinite(range)) {
-        error->all(FLERR, "interaction_range is infinite for this model, this is not yet supported");
-    } else {
-        this->interaction_range = range;
-    }
 
     // get the model's preferred devices
     bool found_valid_device = false;
@@ -388,10 +371,6 @@ void PairMetatensor::load_torch_model(const char* path) {
             path, torch::str(this->capabilities->supported_devices)
         );
     }
-
-    // move all data to the right device
-    this->selected_atoms_values = selected_atoms_values.to(device);
-    this->atomic_types = atomic_types.to(device);
 
     if (lmp->comm->me == 0) {
         auto metadata_ivalue = this->torch_model->run_method("metadata");
@@ -412,312 +391,4 @@ void PairMetatensor::load_torch_model(const char* path) {
             }
         }
     }
-}
-
-static std::array<int32_t, 3> cell_shifts(
-    const std::array<std::array<double, 3>, 3>& cell_inv,
-    const std::array<double, 3>& pair_shift
-) {
-    auto shift_a = static_cast<int32_t>(std::round(
-        cell_inv[0][0] * pair_shift[0] +
-        cell_inv[0][1] * pair_shift[1] +
-        cell_inv[0][2] * pair_shift[2]
-    ));
-    auto shift_b = static_cast<int32_t>(std::round(
-        cell_inv[1][0] * pair_shift[0] +
-        cell_inv[1][1] * pair_shift[1] +
-        cell_inv[1][2] * pair_shift[2]
-    ));
-    auto shift_c = static_cast<int32_t>(std::round(
-        cell_inv[2][0] * pair_shift[0] +
-        cell_inv[2][1] * pair_shift[1] +
-        cell_inv[2][2] * pair_shift[2]
-    ));
-
-    return {shift_a, shift_b, shift_c};
-}
-
-
-void PairMetatensor::setup_neighbors(metatensor_torch::System& system) {
-    auto dtype = system->positions().scalar_type();
-
-    double** x = atom->x;
-    auto total_n_atoms = atom->nlocal + atom->nghost;
-
-    auto cell_inv_tensor = system->cell().inverse().t();
-    auto cell_inv_accessor = cell_inv_tensor.accessor<double, 2>();
-    auto cell_inv = std::array<std::array<double, 3>, 3>{{
-        {{cell_inv_accessor[0][0], cell_inv_accessor[0][1], cell_inv_accessor[0][2]}},
-        {{cell_inv_accessor[1][0], cell_inv_accessor[1][1], cell_inv_accessor[1][2]}},
-        {{cell_inv_accessor[2][0], cell_inv_accessor[2][1], cell_inv_accessor[2][2]}},
-    }};
-
-    // Collect the local atom id of all local & ghosts atoms, mapping ghosts
-    // atoms which are periodic images of local atoms back to the local atoms.
-    //
-    // Metatensor expects pairs corresponding to periodic atoms to be between
-    // the main atoms, but using the actual distance vector between the atom and
-    // the ghost.
-    auto original_atom_id = std::vector<int>();
-    original_atom_id.reserve(total_n_atoms);
-
-    // identify all local atom by their LAMMPS atom tag.
-    // TODO: cache this allocation
-    auto local_atoms_tags = std::unordered_map<tagint, int>();
-    for (int i=0; i<atom->nlocal; i++) {
-        original_atom_id.emplace_back(i);
-        local_atoms_tags.emplace(atom->tag[i], i);
-    }
-
-    // now loop over ghosts & map them back to the main cell if needed
-    auto ghost_atoms_tags = std::unordered_map<tagint, int>();
-    for (int i=atom->nlocal; i<total_n_atoms; i++) {
-        auto tag = atom->tag[i];
-        auto it = local_atoms_tags.find(tag);
-        if (it != local_atoms_tags.end()) {
-            // this is the periodic image of an atom already owned by this domain
-            original_atom_id.emplace_back(it->second);
-        } else {
-            // this can either be a periodic image of an atom owned by another
-            // domain, or directly an atom from another domain. Since we can not
-            // really distinguish between these, we take the first atom as the
-            // "main" one and remap all atoms with the same tag to the first one
-            auto it = ghost_atoms_tags.find(tag);
-            if (it != ghost_atoms_tags.end()) {
-                // we already found this atom elsewhere in the system
-                original_atom_id.emplace_back(it->second);
-            } else {
-                // this is the first time we are seeing this atom
-                original_atom_id.emplace_back(i);
-                ghost_atoms_tags.emplace(tag, i);
-            }
-        }
-    }
-
-    for (auto& cache: this->neigh_cache) {
-        auto cutoff2 = cache.cutoff * cache.cutoff;
-        auto full_list = cache.options->full_list();
-
-        // convert from LAMMPS neighbors list to metatensor format
-        cache.known_samples.clear();
-        cache.samples.clear();
-        cache.distances_f32.clear();
-        cache.distances_f64.clear();
-        for (int ii=0; ii<(list->inum + list->gnum); ii++) {
-            auto atom_i = list->ilist[ii];
-            auto original_atom_i = original_atom_id[atom_i];
-
-            auto neighbors = list->firstneigh[ii];
-            for (int jj=0; jj<list->numneigh[ii]; jj++) {
-                auto atom_j = neighbors[jj];
-                auto original_atom_j = original_atom_id[atom_j];
-
-                if (!full_list && original_atom_i > original_atom_j) {
-                    // Remove extra pairs if the model requested half-lists
-                    continue;
-                }
-
-                auto distance = std::array<double, 3>{
-                    x[atom_j][0] - x[atom_i][0],
-                    x[atom_j][1] - x[atom_i][1],
-                    x[atom_j][2] - x[atom_i][2],
-                };
-
-                auto distance2 = (
-                    distance[0] * distance[0] +
-                    distance[1] * distance[1] +
-                    distance[2] * distance[2]
-                );
-                if (distance2 > cutoff2) {
-                    // LAMMPS neighbors list contains some pairs after the
-                    // cutoff, we filter them here
-                    continue;
-                }
-
-                // Compute the cell shift for the pair.
-                auto shift_i = std::array<double, 3>{
-                    x[atom_i][0] - x[original_atom_i][0],
-                    x[atom_i][1] - x[original_atom_i][1],
-                    x[atom_i][2] - x[original_atom_i][2],
-                };
-                auto shift_j = std::array<double, 3>{
-                    x[atom_j][0] - x[original_atom_j][0],
-                    x[atom_j][1] - x[original_atom_j][1],
-                    x[atom_j][2] - x[original_atom_j][2],
-                };
-                auto pair_shift = std::array<double, 3>{
-                    shift_j[0] - shift_i[0],
-                    shift_j[1] - shift_i[1],
-                    shift_j[2] - shift_i[2],
-                };
-
-                auto shift = std::array<int32_t, 3>{0, 0, 0};
-                if (pair_shift[0] != 0 || pair_shift[1] != 0 || pair_shift[2] != 0) {
-                    shift = cell_shifts(cell_inv, pair_shift);
-
-                    if (!full_list && original_atom_i == original_atom_j) {
-                        // If a half neighbors list has been requested, do
-                        // not include the same pair between an atom and
-                        // it's periodic image twice with opposite cell
-                        // shifts (e.g. [1, -1, 1] and [-1, 1, -1]).
-                        //
-                        // Instead we pick pairs in the positive plan of
-                        // shifts.
-                        if (shift[0] + shift[1] + shift[2] < 0) {
-                            // drop shifts on the negative half-space
-                            continue;
-                        }
-
-                        if ((shift[0] + shift[1] + shift[2] == 0)
-                            && (shift[2] < 0 || (shift[2] == 0 && shift[1] < 0))) {
-                            // drop shifts in the negative half plane or the
-                            // negative shift[1] axis.
-                            //
-                            // See below for a graphical representation: we are
-                            // keeping the shifts indicated with `O` and
-                            // dropping the ones indicated with `X`
-                            //
-                            //  O O O │ O O O
-                            //  O O O │ O O O
-                            //  O O O │ O O O
-                            // ─X─X─X─┼─O─O─O─
-                            //  X X X │ X X X
-                            //  X X X │ X X X
-                            //  X X X │ X X X
-                            continue;
-                        }
-                    }
-                }
-
-                auto sample = std::array<int32_t, 5>{
-                    original_atom_i,
-                    original_atom_j,
-                    shift[0],
-                    shift[1],
-                    shift[2],
-                };
-
-                // only add the pair if it is not already known. The same pair
-                // can occur multiple time between two periodic ghosts shifted
-                // around by the same amount, but we only want one of these pairs.
-                if (cache.known_samples.insert(sample).second) {
-                    cache.samples.push_back(sample);
-
-                    if (dtype == torch::kFloat64) {
-                        cache.distances_f64.push_back(distance);
-                    } else if (dtype == torch::kFloat32) {
-                        cache.distances_f32.push_back({
-                            static_cast<float>(distance[0]),
-                            static_cast<float>(distance[1]),
-                            static_cast<float>(distance[2])
-                        });
-                    } else {
-                        // should be unreachable
-                        error->all(FLERR, "invalid dtype, this is a bug");
-                    }
-                }
-            }
-        }
-
-        int64_t n_pairs = cache.samples.size();
-        auto samples_values = torch::from_blob(
-            reinterpret_cast<int32_t*>(cache.samples.data()),
-            {n_pairs, 5},
-            torch::TensorOptions().dtype(torch::kInt32).device(torch::kCPU)
-        );
-
-        auto samples = torch::make_intrusive<metatensor_torch::LabelsHolder>(
-            std::vector<std::string>{"first_atom", "second_atom", "cell_shift_a", "cell_shift_b", "cell_shift_c"},
-            samples_values
-        );
-
-        auto distances_vectors = torch::Tensor();
-        if (dtype == torch::kFloat64) {
-            distances_vectors = torch::from_blob(
-                cache.distances_f64.data(),
-                {n_pairs, 3, 1},
-                torch::TensorOptions().dtype(torch::kFloat64).device(torch::kCPU)
-            );
-        } else if (dtype == torch::kFloat32) {
-            distances_vectors = torch::from_blob(
-                cache.distances_f32.data(),
-                {n_pairs, 3, 1},
-                torch::TensorOptions().dtype(torch::kFloat32).device(torch::kCPU)
-            );
-        } else {
-            // should be unreachable
-            error->all(FLERR, "invalid dtype, this is a bug");
-        }
-
-        auto neighbors = torch::make_intrusive<metatensor_torch::TensorBlockHolder>(
-            distances_vectors,
-            samples,
-            std::vector<metatensor_torch::TorchLabels>{
-                metatensor_torch::LabelsHolder::create({"xyz"}, {{0}, {1}, {2}}),
-            },
-            metatensor_torch::LabelsHolder::create({"distance"}, {{0}})
-        );
-
-        metatensor_torch::register_autograd_neighbors(system, neighbors, this->check_consistency);
-        system->add_neighbors_list(cache.options, neighbors);
-    }
-}
-
-
-metatensor_torch::System PairMetatensor::system_from_lmp(bool do_virial) {
-    auto dtype = torch::kFloat64;
-    if (this->capabilities->dtype() == "float64") {
-        dtype = torch::kFloat64;
-    } else if (this->capabilities->dtype() == "float32") {
-        dtype = torch::kFloat64;
-    } else {
-        error->all(FLERR, "the model requested an unsupported dtype '{}'", this->capabilities->dtype());
-    }
-
-    double** x = atom->x;
-    auto total_n_atoms = atom->nlocal + atom->nghost;
-
-    this->atomic_types.resize_({total_n_atoms});
-    for (int i=0; i<total_n_atoms; i++) {
-        this->atomic_types[i] = this->lammps_to_metatensor_types[atom->type[i]];
-    }
-
-    auto tensor_options = torch::TensorOptions().dtype(torch::kFloat64).device(torch::kCPU);
-
-    // atom->x contains "real" and then ghost atoms, in that order
-    auto positions = torch::from_blob(
-        *x, {total_n_atoms, 3},
-        // requires_grad=true since we always need gradients w.r.t. positions
-        tensor_options.requires_grad(true)
-    );
-
-    auto cell = torch::zeros({3, 3}, tensor_options);
-    cell[0][0] = domain->xprd;
-
-    cell[1][0] = domain->xy;
-    cell[1][1] = domain->yprd;
-
-    cell[2][0] = domain->xz;
-    cell[2][1] = domain->yz;
-    cell[2][2] = domain->zprd;
-
-    if (do_virial) {
-        this->strain = torch::eye(3, tensor_options.requires_grad(true));
-
-        // pretend to scale positions/cell by the strain so that it enters the
-        // computational graph.
-        positions = positions.matmul(this->strain);
-        positions.retain_grad();
-
-        cell = cell.matmul(this->strain);
-    }
-
-    auto system = torch::make_intrusive<metatensor_torch::SystemHolder>(
-        atomic_types,
-        positions.to(dtype),
-        cell.to(dtype)
-    );
-
-    this->setup_neighbors(system);
-    return system;
 }
